Make bit ops use `u32` for shifts and bit counts (#373)

+++ src/macros.rs
@@ -59,7 +59,7 @@ macro_rules! const_assert_ne {
 #[macro_export]
 macro_rules! nlimbs {
     ($bits:expr) => {
-        $bits / $crate::Limb::BITS
+        $bits / $crate::Limb::BITS as usize
     };
 }
